Refactor runtime to use libraries for each platform
